Split fake RP into its own process
