+++ .gitignore
@@ -14,3 +14,4 @@
 .idea
 coverage.out
 /secrets/*
+/fakerp
