+++ pkg/fakerp/config.go
@@ -5,77 +5,17 @@ import (
 	"crypto/x509"
 	"fmt"
 	"io/ioutil"
-	"math/rand"
 	"os"
 	"path/filepath"
-	"time"
-
-	"github.com/kelseyhightower/envconfig"
-	"github.com/sirupsen/logrus"
 
 	"github.com/openshift/openshift-azure/pkg/api"
+	"github.com/openshift/openshift-azure/pkg/fakerp/shared"
 	"github.com/openshift/openshift-azure/pkg/tls"
 )
 
-var supportedRegions = []string{
-	"australiasoutheast",
-	"eastus",
-	"westeurope",
-}
-
-type Config struct {
-	Username        string `envconfig:"AZURE_USERNAME"`
-	Password        string `envconfig:"AZURE_PASSWORD"`
-	SubscriptionID  string `envconfig:"AZURE_SUBSCRIPTION_ID" required:"true"`
-	TenantID        string `envconfig:"AZURE_TENANT_ID" required:"true"`
-	ClientID        string `envconfig:"AZURE_CLIENT_ID" required:"true"`
-	ClientSecret    string `envconfig:"AZURE_CLIENT_SECRET" required:"true"`
-	AADClientID     string `envconfig:"AZURE_AAD_CLIENT_ID"`
-	AADClientSecret string `envconfig:"AZURE_AAD_CLIENT_SECRET"`
-
-	Region           string `envconfig:"AZURE_REGION"`
-	DnsDomain        string `envconfig:"DNS_DOMAIN"`
-	DnsResourceGroup string `envconfig:"DNS_RESOURCEGROUP"`
-	ResourceGroup    string `envconfig:"RESOURCEGROUP" required:"true"`
-
-	NoGroupTags      bool   `envconfig:"NOGROUPTAGS"`
-	ResourceGroupTTL string `envconfig:"RESOURCEGROUP_TTL"`
-	Manifest         string `envconfig:"MANIFEST"`
-	NoWait           bool   `envconfig:"NO_WAIT"`
-}
-
 const (
-	DataDirectory           = "_data/"
-	LocalHttpAddr           = "localhost:8080"
 	LoggingSecretsDirectory = "secrets/"
 )
-
-func NewConfig(log *logrus.Entry, needRegion bool) (*Config, error) {
-	var c Config
-	if err := envconfig.Process("", &c); err != nil {
-		return nil, err
-	}
-	if needRegion {
-		if c.Region == "" {
-			// Randomly assign a supported region
-			rand.Seed(time.Now().UTC().UnixNano())
-			c.Region = supportedRegions[rand.Intn(len(supportedRegions))]
-			log.Infof("using randomly selected region %s", c.Region)
-		}
-
-		var supported bool
-		for _, region := range supportedRegions {
-			if c.Region == region {
-				supported = true
-			}
-		}
-		if !supported {
-			return nil, fmt.Errorf("%s is not a supported region (supported regions: %v)", c.Region, supportedRegions)
-		}
-		os.Setenv("AZURE_REGION", c.Region)
-	}
-	return &c, nil
-}
 
 func GetPluginConfig() (*api.PluginConfig, error) {
 	tc := api.TestConfig{
@@ -91,7 +31,7 @@ func GetPluginConfig() (*api.PluginConfig, error) {
 	}
 
 	// populate geneva artifacts
-	artifactDir, err := FindDirectory(LoggingSecretsDirectory)
+	artifactDir, err := shared.FindDirectory(LoggingSecretsDirectory)
 	if err != nil {
 		return nil, err
 	}
