+++ test/clients/openshift/login.go
@@ -12,7 +12,7 @@ import (
 	"k8s.io/client-go/tools/clientcmd/api"
 	"k8s.io/client-go/tools/clientcmd/api/latest"
 
-	"github.com/openshift/openshift-azure/pkg/fakerp"
+	fakerp "github.com/openshift/openshift-azure/pkg/fakerp/shared"
 	azuretls "github.com/openshift/openshift-azure/pkg/tls"
 	"github.com/openshift/openshift-azure/pkg/util/managedcluster"
 )
